[pp] Minor optimization in `Pp.t` construction and gluing.
The typical Coq `Pp.t` document contains a lot of "gluing" which
produces efficient structures but it is quite painful in
serialization.

We optimize a common document building case so we don't create as much
glue nodes as with the "naive" strategy, and without incurring in the
large performance cost full flattening would produce.

This is a temporal fixup, see #505 for more context on the discussion
and medium-term plans.

(cherry picked from commit 8fa1d224dafde29f5b527d380c069f196a042c60)