Filters: remove from evar_info
This commit entirely removes evar_filter from evar_info, and accordingly
simplifies the API of evar_maps, with just one function to get the
hyps/env/etc of an evar. Filters are still used to conduct pruning
throughout unification/case compilation etc, but the restriction of an
evar by a filter is always done immediately.