Remove "API/PROPERTIES" file.
There are two problems with this file:
- at the moment, it is out of date
- without API/ROADMAP, it does not make much sense.

I propose to remove it for now and add it only with a roadmap.
The initial contents could be:

--------------------------------------------------------------------------------
All the (currently maintained) API properties are listed below.
If you notice that API deviates from them, it is considered as a bug
and that fact should be reported to Coq's bugzilla.

0 : all plugins, that are distributed with Coq, are compilable
    (without "-bypass-API" option)
--------------------------------------------------------------------------------