Document the known problems of the Coq API.
We know that Coq API is imperfect (to say the least).
It has many problems.
Some of them simpler, some of them more serious.
Some of them easier to fix, some of them harder to fix.
Some of them more and some of them less relevant (for plugin writers).

While we could simply keep a list of problems
(either in a text file, or in a bugzilla)
we would quickly realize that it is hard to avoid adding duplicate entries for the same problem(s).
We want to avoid that because that would increase the noise.

Also, if we tried simply to keep a list of problems
(either in a text file, or in a bugzilla)
it would be difficult for us to see "all problems of the same particular kind".
(if we used Jira instead of Bugzilla, then we might be able to use tags to address this problem)

Also, it appears that it is possible to define "kinds of problems",
it is natural to try to organize them hierarchically (if possible)
instead of having a linear list of "kinds of problems".

I am not sure what tool could be collectively used to maintain such information
(about "kinds of problems" as well as their particular "instances").
Originally, I maintained this information in my personal mindmap.
That worked, but the drawback were that:
- it was not possible to share
- it was outside of the versioning system
The simplest solution I was able to find was to convert the original information
stored in the mindmap into directories and files.

Each file represents a single "kind of a problem".
The names of files are supposed to be the names of "kinds of problems".
Each file starts with a general description of a given "kind of a problem".
This is followed by a vertical line: ===================================== ... =
which is followed by a list of "instances of a given problem".
When I knew thatelipsis at the list of instances.

TODO:
- identify more "kinds of problems"
  (this may result in restructuring the hierarchy of "kinds of problems")
- identify more "instances" of the known "kinds of problems"

MOTIVATION:
- If we want to define a roadmap, we need to be aware of the problems that has to be solved
  (before we can hope to deliver an API that is not shocking).