Better error messages on the CI
This makes it so that when a ci target fails, we don't get red herring
error messages about .ok files not existing.

Since this requires bash, we make a helper script that invokes bash, so
as to not depend on bash in general.

(cherry picked from commit 23d1504de4111223dc4805c6f4176a4d9e028b79)