equality: port rewrite, injection and discriminate
Support for patterns attached to lemmas, abstract all occurrences of the
lhs and proof in rewrite.  Working with occurrences, not relying on
setoid anymore

- injection with_evars flag, using clenv
- default inferred pattern is more approximate
  Does not index lambdas or any implicit arguments (they get replaced
  by _).
  simplify inferred filter: head + application structure

  Also ensure that arguments are converted independently so that the
  unifier matches the application structure. This is pretty permissive but
  seems to preserve compatibility of proof scripts well.

- refine test for patterns, configuration of indexes

  Take into account the transparent state of "rewrite" hint db for
indexing and unification. Now by default allowing to unfold local
variables during unification. Rrespect the complete application
structure of the inferred pat. This allows to rewrite with a lemma
like (lem (_ _)) ensuring we find occurrences of the explicit
application using first-order unification as long as it is under an
indexed part of the lemma.

- index pattern on non-evaluable constants
- flags for rewrite selection of subterms
- frozen_evars for rewrite
- simplify instantiate_lemma
- keyed unification heuristic
- use modulo_betaiota to do two phases: first path selecting occurrences
without reduction, second path with reduction
- fix discriminate on on-the-fly ind principle

- rewrite.ml: remove redefinition of "rewrite" hintdb