[coq_makefile] Enable `-safe-string` on plugin building.
Unfortunately it seems that merging `coq_makefile` 2 conflicted with
the merge of the `-safe-string` branch, and plugins are not build with
`-safe-string` anymore.

This is a serious bug, we re-enable the flag again, factorizing some
flag code in the way. [More should be done thou]

We also remove an outdated comment.