Restoring test on ident validity while browsing directory structure.
The test was abandoned at the time of merging subdirectory browsing
between coqdep and coqtop, and to limit at the same time the
dependency of coqdep in files such as unicode.cmo.

But checking ident validity speeds up browsing in arbitrary directory
structure and we restore it for this reason.

(One could also say that browsing arbitrary directory structures is
not intended, but in practice this may happen, as e.g. reported in
BZ#5734.)