[stm] Make the main interfaces functional.
We make the main STM interfaces functional, removing most state from
`stm.ml`.

In particular, error handling has changed quite a lot, as we must not
restore the document. Further improvements are possible in this area.

With regards to the worker interface, we include the document in the
functional task, this should be OK for now but in the future we may
need to use more refined rules.

After some discussion with @gares we have also removed the "keep
states" optimization in the proof worker. It should be introduced in
the future once we make the TaskQueue API aware of worker state, but
IMHO TaskQueue improvements should go in a separate commit.

Be should be careful to check if there is any performance impact thou.