[general] Move declare to pretyping.
This PR is a proof of concept following discussion with @herbelin;
more refinement would be needed. We basically move all the
functionality specific to interpretation to `Ideclare`.

The idea is for every module provide an enriched "declare" function,
so users above it can choose to declare using that functionality.

I think the use of schemes should be fixed too, instead of using a
hook.

Note that vernac already has declareDef, maybe some of the
functionality should be mixed together.

Indeed, this belongs to a more general discussion on how to extend the
functionality of core components; maybe could be a good topic for the WG.