Moving setting of "cleared" evar flag directly in Evd.restrict.
In particular, this fixes #5757 which used restrict_evar to refine the
information on the source of an evar, and which should have set the
"cleared" flag.

Also renaming flag "restricted" since it is not only about "clear".

I guess this is what we want in general, but I did not survey all uses
of restrict_evar so, maybe, this should be refined further.

(cherry picked from commit 05bd0ab1dd85764874ca077005dcaff5414589a5)