[ltac] Remove Deprecated `Add Morphism` form.
We remove deprecated form `Add Morphism op : id`.

The reason is that this is one of the few commands to require
"dynamic" classification; c.f. #459 #566 #1095, which is quite
problematic.

Another option to keep the command would be to require that it always
starts a proof; however, as the command is marked as deprecated we try
its removal with the consequent cleanup.

Note that the documentation marks the command as always starting a
proof, however I deduced from the classifier history that this is not
correct.