[ltac] Warning for deprecated `Add Setoid` and `Add Morphism` forms.
The manual has long stated that these forms are deprecated. We add a
warning for them, as indeed `Add Morphism` is an "proof evil" [*]
command, and we may want to remove it in the future.

We've also fixed the stdlib not to emit the warning.

[*] https://ncatlab.org/nlab/show/principle+of+equivalence

(cherry picked from commit 6526933e3d6a6392aa6bd5235ea0180bb68b6f94)