[stm] Remove Opacity Information from Theorem and Definition start.
This information was introduced in
db65876404c7c3a1343623cc9b4d6c2a7164dd95 and IMHO was wrongly used all
the way around.

It seems that some features of the proof layer and the document/STM
ought to be better integrated; in particular as discussed in #682 we
may want to revert eef907ed0a200e912ab2eddc0fcea41b5f61c145 and in
fact go back to a design where the proof terms a proof is allowed to
produce correspond to the information stored in `VtStartProof`.

I agree with the spirit of db65876404c7c3a1343623cc9b4d6c2a7164dd95 in
the sense that the current static analysis is tied to document
traversal and this is inflexible in some scenarios; however we may
want to be more regular at the vernac level too.

This PR is more a RFC than a request for immediate merge.