An attempt to fix issue #5771 (error color hidden by warning color).
We change the relative priority of errors and warnings, so that the
error takes precedence.

It is unsure that it is universally the best choice. If the location
of the error is finer than the one of the warning, it is better. In
the other way round, it might be less good, e.g. if understanding the
warning helps to understand the error.

Maybe the best policy would be to test the relative locations of the
warning and error?

Trying to consider the error as more important, at the current time.

(cherry picked from commit d651d81a13d78f0466d2a08c06e7f2318c42da5f)