Ltac Debug: exporting env and sigma when needed so that term can be printed.
We do it so as to preserve non-focussing semantics for non-focussing
generic arguments.

This assumes that the code treats them consistently, which is not
enforced statically, but which is reasonable in the sense that when we
need a context for printing, we have no other choice as needing a
context and we needed one also at interpretation time.