[ci] [coq] Complete 4.06.0 support.
Due to an API change in laglgtk, we need to update CoqIDE. We use a
makefile hack so it can compile with lablgtk < 2.8.16, another option
would be to require 2.8.16 as a minimal dependency.

We also refactor travis to test more lablgtk versions.

We also need to account for improved attribute handling in 4.06.0, in
particular module aliases will propagate the deprecation status.

Fixes #6140.