[api] Move `Names` out of API.
This is a proof of concept in order to do the third iteration of the
"API" design.

The idea is provide in each module needing it a `Public` and
`Internal` system. Then, we make the public part available to
consumers.

This provides significant advantages wrt previous approach, at a
non-so-modest cost of having to update the base implementation
itself. However, the transition could be implemented gradually.

c.f. issue #6008