Fixing encoding in coqdoc output tests.
The file links.v is using utf-8 characters so this is needed at least
to test this file. For the other files, it is not completely without
effect since it makes that symbols like => and forall are respectively
displayed ⇒ and ∀.

Maybe tests with iso-8859-1 or test without a charset option should be
kept.