Ssreflect.v: drop cast notation (see #6078 for discussion).
The effect of the notation on the grammar and printing was not
different from what was done in Coq (since 8.5).

However, it had the effect of blocking the level for ":" what created
incompatibilities with other user choices (even though these other
uses are themselves fragile since they modify a level hard-wired in
the grammar of terms).