[api] Move `global_reference` to Kernel.Names.
This was suggested initially by Gaëtan Gilbert in #6092, however it
makes sense to go further and use the current kernel style for names.

This relocation allows to improve the ordering of API significantly
as some core modules don't depend on library anymore.

The main question here is about modularization of the global_reference
stuff itself, given that we have two different notions of equality for
constants. In that sense, the KerPair is suspicious and at some point
it should be looked at.