Use a more compact representation for bytecode relocations stored on disk.
The previous implementation used a list of pairs, which has size 9n where n
is the number of relocations. We instead use two arrays for a total memory
cost of 2n + 5 words.

The use of arrays may turn out to be problematic on 32-bit machines, I am unsure
if we will hit this limitation in practice.