Forbid implicitly relative names in the checker.
Before this patch, passing a mere identifier (without dots) to the checker
would make it consider it as implicitly referring to a relative name. For
instance, if passed "foo", it would have looked for "Bar.foo.vo" and
"Qux.foo.vo" if those files were in the loadpath.

This was quite ad-hoc. We remove this "feature" and require the user to
always give either a filename or a fully qualified logical name.