Proof engine: better support for nesting tactic-in-term within other tactics.
the possibility of saving previous goals when calling

Tactic-in-term can be called from within a tactic itself. We have to
preserve the preexisting future_goals (if called from pretyping) and
we have to inform of the existence of pending goals, using
future_goals which is the only way to tell it in the absence of being
part of an encapsulating proofview.

This fixes #6313.

Note: we fail if the shelf or given_up are not empty as we don't know
how to preserve their status.