Port the ssr tactic API to the new engine.
In practice, this means that all ssreflect tactics now export new tactic
types. Internally, they still rely on the compatibility layers, but this
means it will be easier to port the code piecewise.

Specific types that were not tactics, like a few pf_* variants, were left
as-is. Still, they should also be given in terms of tactics at some point
as well.