[makefile] Address #6291: install more development files.
As noted in the bug #6291, `.cmx` files are not installed in the
coqlib, which yields the warning:

```
Warning 58: no cmx file was found in path for module Sorts, and its
interface was not compiled with -opaque
```

We thus install the `cmx` files to fix this problem, and also install
the `.o` files for plugins' `.o` to support linking the plugins
statically.

This closes #5099 and #6291.