Add named timers to LtacProf
I'm not sure if they belong in profile_ltac, or in extratactics, or,
perhaps, in a separate plugin.  But I'd find it very useful to have a
version of `time` that works on constr evaluation, which is what this
commit provides.

I'm not sure that I've picked good naming conventions for the tactics,
either.